refactor utils to avoid circular dependency
