+++ data_access_service/batch/subsetting/tasks/generate_dataset.py
@@ -23,10 +23,12 @@ from data_access_service.models.subset_request import NON_SPECIFIED, SubsetReque
 from data_access_service.batch.subsetting.helpers.data_file_upload import (
     upload_all_files_in_folder_to_temp_s3,
 )
+from data_access_service.batch.subsetting.helpers.parquet_date_ranges import (
+    check_rows_with_date_range,
+    trim_date_range,
+)
 from data_access_service.utils.date_time_utils import (
     get_monthly_utc_date_range_array_from_,
-    trim_date_range,
-    check_rows_with_date_range,
 )
 from data_access_service.utils.multi_polygon_helper import merge_polygons
