+++ data_access_service/core/middleware.py
@@ -6,7 +6,7 @@ from starlette.middleware.gzip import GZipMiddleware
 def configure_gzip_middleware(app: FastAPI) -> None:
     """PNG/WebP tiles are already compressed; gzipping them wastes CPU. Core JSON
     responses that already set their own Content-Encoding (see
-    utils/routes_helper.py's gzip_compress) are left untouched by this
+    core/routes/helpers.py's gzip_compress) are left untouched by this
     middleware — it skips compression when content-encoding is already set.
     """
     if "image/" not in _gzip_mw.DEFAULT_EXCLUDED_CONTENT_TYPES:
