+++ data_access_service/core/routes/pmtiles.py
@@ -6,9 +6,9 @@ from data_access_service.batch.pmtiles.generator import (
     PmtilesGenerationInProgressError,
     generate_pmtiles_for_parquets,
 )
-from data_access_service.utils.api_utils import api_key_auth
+from data_access_service.core.routes.auth import api_key_auth
 from data_access_service.utils.date_time_utils import time_it
-from data_access_service.utils.routes_helper import get_api_instance
+from data_access_service.core.routes.helpers import get_api_instance
 from data_access_service.utils.sse_utils import sse_it
 
 router = APIRouter()
