--- data_access_service/utils/stop_watch.py
@@ -1,29 +1,0 @@
-import time
-
-
-class StopWatch:
-    """
-    A simple stopwatch utility to measure the time taken for tasks. It is useful for performance monitoring and debugging.
-    """
-
-    def __init__(self):
-        self.__start_time = None
-        self.__task_name = None
-
-    def start(self, task_name: str):
-        self.__start_time = time.time()
-        self.__task_name = task_name
-
-    def record(self, record_label: str):
-        if self.__start_time is None:
-            raise ValueError("Stopwatch has not been started.")
-        elapsed_time = time.time() - self.__start_time
-        print(f"Timer result: {record_label} took {elapsed_time:.2f} seconds")
-
-    def stop(self):
-        if self.__start_time is None:
-            raise ValueError("Stopwatch has not been started.")
-        elapsed_time = time.time() - self.__start_time
-        print(f"Timer result: {self.__task_name} took {elapsed_time:.2f} seconds")
-        self.__start_time = None
-        self.__task_name = None
